Extract test helpers to a new base class
